@@ -3,21 +3,22 @@
   "version": "1.0.0",
   "license": "MIT",
   "scripts": {
-    "build": "rollup --config rollup.config.js --environment BUILD:production",
-    "test": "yarn build && mkdir -p test/empty_vault/.obsidian/plugins/sample-plugin && cp manifest.json styles.css main.js test/empty_vault/.obsidian/plugins/sample-plugin && mocha --exit"
+    "build": "rollup --config rollup.config.mjs --environment BUILD:production",
+    "package": "yarn build && mkdir -p test/empty_vault/.obsidian/plugins/sample-plugin && cp manifest.json styles.css dist/main.js test/empty_vault/.obsidian/plugins/sample-plugin",
+    "test": "yarn package && mocha --exit"
   },
   "packageManager": "yarn@4.2.1+sha256.15ce76682a8cd2090257b883cd69c637925b29573f9573e8403ec227d5ab6815",
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^18.0.0",
-    "@rollup/plugin-node-resolve": "^11.2.1",
-    "@rollup/plugin-typescript": "^8.2.1",
-    "@types/node": "^14.14.37",
-    "chai": "^4.3.4",
-    "mocha": "^8.3.2",
-    "obsidian": "^0.11.13",
-    "rollup": "^2.32.1",
-    "spectron": "13",
-    "tslib": "^2.2.0",
-    "typescript": "^4.2.4"
+    "@rollup/plugin-commonjs": "^25.0.7",
+    "@rollup/plugin-node-resolve": "^15.2.3",
+    "@rollup/plugin-typescript": "^11.1.6",
+    "@types/node": "^20.12.10",
+    "chai": "^5.1.0",
+    "mocha": "^10.4.0",
+    "obsidian": "^1.5.7",
+    "rollup": "^4.17.2",
+    "spectron": "19",
+    "tslib": "^2.6.2",
+    "typescript": "^5.4.5"
   }
 }
